Updated Streams and FileStreamPeers to use 64-bit addressing.

+++ vcf/src/vcf/FoundationKit/BasicInputStream.h
@@ -47,7 +47,7 @@ public:
 
 	virtual uint64 getCurrentSeekPos() ;
 
-	virtual uint32 read( unsigned char* bytesToRead, uint32 sizeOfBytes );
+	virtual uint64 read( unsigned char* bytesToRead, uint64 sizeOfBytes );
 
 	virtual bool isEOS();
